Fixing usage for getInteger which should return an int
git-svn-id: https://www.silverpeas.org/svn/silverpeas/core/trunk@3020 a8e77078-a1c7-4fa5-b8fc-53c5178a176c

+++ lib-core/src/main/java/com/stratelia/webactiv/util/ResourceLocator.java
@@ -241,7 +241,7 @@ public class ResourceLocator implements Serializable {
    * bundle.
    * @return the value as an integer.
    */
-  public long getInteger(final String sAttribute, long defaultValue) {
+  public int getInteger(final String sAttribute, int defaultValue) {
     String value = getString(sAttribute);
     if (value == null || value.trim().isEmpty()) {
       return defaultValue;
